upgrade client

+++ src/main/java/org/petos/packagemanager/client/PackageInstaller.java
@@ -1,6 +1,5 @@
 package org.petos.packagemanager.client;
 
-import com.google.gson.Gson;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
@@ -8,11 +7,11 @@ import org.petos.packagemanager.packages.FullPackageInfoDTO;
 import org.petos.packagemanager.packages.PackageAssembly;
 
 import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.nio.file.attribute.PosixFilePermission;
 import java.util.ArrayList;
+import java.util.EnumSet;
 import java.util.List;
 
 /**
@@ -20,7 +19,6 @@ import java.util.List;
  */
 public class PackageInstaller {
 private static final Logger logger = LogManager.getLogger(PackageInstaller.class);
-
 public static class PackageIntegrityException extends Exception {
       PackageIntegrityException(String msg) {
 	    super(msg);
@@ -37,8 +35,7 @@ public enum PayloadType {Binary, Library, Docs, Config, Unknown}
 
 public enum CommitState {Failed, Success}
 
-;
-private Configuration config;
+private final Configuration config;
 
 public PackageInstaller(Configuration config) {
       this.config = config;
@@ -48,29 +45,68 @@ public void initSession() {
       centralPath = null;
       dependencies = new ArrayList<>();
       try {
-	    tempConfig = Files.createTempFile(Path.of(getTempPath()), "config", "pum");
+	    tempConfig = Path.of(getTempPath(), "config.pum");
+	    Files.deleteIfExists(tempConfig);
+	    Files.createFile(tempConfig);
       } catch (IOException e) {
 	    //really interesting case...
 	    throw new RuntimeException(e);
       }
 }
 
+/**
+ * Store package in local file system
+ * and add package's info to local registry<br>
+ * The following methods are fully depends on OS. As rule, the following sample is very primitive and generalized
+ */
+public void storeLocally(FullPackageInfoDTO dto, PackageAssembly assembly) throws PackageIntegrityException {
+      PayloadType type = convert(dto.payloadType);
+      if (type == PayloadType.Unknown)
+	    throw new PackageIntegrityException("Unknown package typ");
+      try {
+	    //the directory where package should be installed -> the method is full safe
+	    Path normalized = normalizePath(dto.name, type);
+	    if (Files.exists(normalized)) { //old version of package should be retired
+		  removeFiles(normalized);
+	    }
+	    Files.createDirectory(normalized);
+	    switch (type) {
+		  case Binary -> {
+			Path binDir = Path.of(normalized.toString(), "bin");
+			Path binaryPath = Path.of(normalized.toString(), "bin", dto.name + ".exe");
+			Files.createDirectory(binDir);
+			Files.write(binaryPath, assembly.getPayload());
+			Files.setPosixFilePermissions(binaryPath, EnumSet.of(PosixFilePermission.OWNER_EXECUTE));
+			addBinary(normalized, assembly.getId(), dto);
+		  }
+		  case Library -> {
+			Path libPath = Path.of(normalized.toString(), dto.name + ".ddl");
+			Files.write(libPath, assembly.getPayload());
+			Files.setPosixFilePermissions(libPath, EnumSet.of(PosixFilePermission.OWNER_EXECUTE)); //or mark as shared library
+			addLibrary(normalized, assembly.getId(), dto);//save installation to regist
+		  }
+	    }
+	    PackageInfo info = PackageStorage.toLocalFormat(assembly, dto);
+	    PackageStorage.storePackageInfo(normalized, info);
+      } catch (IOException e) {
+	    throw new PackageIntegrityException("Payload installation error: " + e.getMessage());
+      }
+}
 public void commitSession(CommitState state) throws PackageIntegrityException {
       checkSession();
       if (centralPath == null)
 	    throw new PackageIntegrityException("Package without binary payload");
       if (state == CommitState.Success) {
 	    try {
-		  byte[] configBytes = Files.readAllBytes(tempConfig);
-		  Path configPath = Path.of(getPackagesConfig());
-		  Files.write(configPath, configBytes, StandardOpenOption.APPEND);
+		  String config = Files.readString(tempConfig);
+		  PackageStorage.rebuildConfig(InstanceInfo.valueOf(config));
 		  Path linkPath = Path.of(centralPath.toString(), "bin", "ddl.conf");
-		  linkLibraries(linkPath, dependencies);
+		  linkLibraries(linkPath, exeFile, dependencies);
 		  clearSession();
 	    } catch (IOException e) {
-		  logger.warn("Package linking error: " + centralPath);
+		  logger.warn("Package linking error: " + e.getMessage());
 		  eraseSession();
-		  throw new PackageIntegrityException("Error during package linking");
+		  throw new PackageIntegrityException("Package linking error" + e.getMessage());
 	    }
       } else {
 	    eraseSession();
@@ -78,8 +114,8 @@ public void commitSession(CommitState state) throws PackageIntegrityException {
 }
 
 private void clearSession() throws IOException {
-      assert tempConfig != null;
-      Files.deleteIfExists(tempConfig);
+      if (tempConfig != null)
+      	Files.deleteIfExists(tempConfig);
       centralPath = null;
       dependencies = null;
       tempConfig = null;
@@ -96,95 +132,56 @@ private void eraseSession() {
 	    }
 	    clearSession();
       } catch (IOException ignored) {
-	    logger.warn("Failed to remove broken package");
+	    logger.warn("Failed to remove broken package: " + ignored.getMessage());
       }
 }
 
-private void saveConfiguration(@NotNull Path installation, @NotNull FullPackageInfoDTO dto) {
+private void saveConfiguration(@NotNull Path installation, @NotNull Integer packageId, @NotNull FullPackageInfoDTO dto) {
       assert dto.aliases != null;
       String[] aliases = new String[dto.aliases.length + 1];
       System.arraycopy(dto.aliases, 0, aliases, 0, dto.aliases.length);
       aliases[dto.aliases.length] = dto.name;
-      appendConfig(tempConfig, aliases, installation);
-
+      appendTempConfig(packageId, aliases, installation);
+}
+private void appendTempConfig(Integer packageId, String[] aliases, Path instancePath) {
+      try {
+	    var info = new InstanceInfo(packageId, aliases, instancePath.toAbsolutePath().toString());
+	    Files.writeString(tempConfig, info.toString(), StandardOpenOption.APPEND);
+      } catch (IOException e) {
+	    throw new RuntimeException("Configuration file is not exists");
+      }
 }
 
 /**
  * Save dependency payload path
  */
-private void addDependency(@NotNull Path installation, @NotNull FullPackageInfoDTO dto) throws PackageIntegrityException {
+private void addLibrary(@NotNull Path installation, @NotNull Integer id, @NotNull FullPackageInfoDTO dto) throws PackageIntegrityException {
       checkSession();
       dependencies.add(installation);
-      saveConfiguration(installation, dto);
+      saveConfiguration(installation, id, dto);
 }
 
-private void addBinary(@NotNull Path packagePath, @NotNull FullPackageInfoDTO dto) throws PackageIntegrityException {
+private void addBinary(@NotNull Path packagePath,@NotNull Integer id, @NotNull FullPackageInfoDTO dto) throws PackageIntegrityException {
       checkSession();
       centralPath = packagePath;
-      saveConfiguration(packagePath, dto);
+      exeFile = centralPath.resolve(Path.of("bin", dto.name + ".exe")).toFile();
+      saveConfiguration(packagePath, id, dto);
 }
 
 private void checkSession() throws PackageIntegrityException {
       if (dependencies == null || tempConfig == null)
 	    throw new PackageIntegrityException("Session is not started");
 }
-
+private File exeFile;
 private Path centralPath; //where binary file is storing
 private Path tempConfig;
 private List<Path> dependencies;
-
-
-/**
- * Store package in local file system
- * and add package's info to local registry<br>
- * The following methods are fully depends on OS. As rule, the following sample is very primitive and generalized
- */
-public void storeLocally(FullPackageInfoDTO info, PackageAssembly assembly) throws PackageIntegrityException {
-      PayloadType type = convert(info.payloadType);
-      if (type == PayloadType.Unknown)
-	    throw new PackageIntegrityException("Unknown package typ");
-      try {
-	    //the directory where package should be installed -> the method is full safe
-	    Path normalized = normalizePath(info.name, type);
-	    if (Files.exists(normalized)) { //old version of package should be retired
-		  removeFiles(normalized);
-		  Files.createDirectory(normalized);
-	    }
-	    switch (type) {
-		  case Binary -> {
-			Path binDir = Path.of(normalized.toString(), "bin");
-			Path binaryPath = Path.of(normalized.toString(), "bin", info.name + ".exe");
-			Files.createDirectory(binDir);
-			Files.write(binaryPath, assembly.getPayload());
-			Files.setAttribute(binaryPath, "unix:+x", true);
-			addBinary(normalized, info);
-		  }
-		  case Library -> {
-			Path libPath = Path.of(normalized.toString(), info.name + ".ddl");
-			Files.write(libPath, assembly.getPayload());
-			Files.setAttribute(normalized, "unix:+x", true);//or mark as shared library
-			addDependency(normalized, info);//save installation to regist
-		  }
-	    }
-	    Path infoPath = Path.of(normalized.toString(), "conf.pum");
-	    byte[] bytesInfo = toByteFormat(assembly, info);
-	    Files.write(infoPath, bytesInfo);
-      } catch (IOException e) {
-	    throw new PackageIntegrityException("Impossible to install package");
-      }
-}
-
-private byte[] toByteFormat(PackageAssembly assembly, FullPackageInfoDTO dto) {
-      String result = String.valueOf(assembly.getId()) + assembly.getVersion() +
-			  new Gson().toJson(dto);
-      return result.getBytes(StandardCharsets.US_ASCII);
-}
 
 private @NotNull Path normalizePath(String name, PayloadType type) {
       Path destPath = null;
       switch (type) {
 	    case Binary -> {
-		  destPath = Path.of(getProgramsPath(), name);
+		  destPath = getProgramsPath().resolve(name);
 	    }
 	    case Library -> {
 		  destPath = Path.of(getLibrariesPath(), name);
@@ -196,38 +193,33 @@ private @NotNull Path normalizePath(String name, PayloadType type) {
 }
 
 private void removeFiles(@NotNull Path filePath) throws IOException {
-      File rootFile = filePath.toFile();
-      if (!rootFile.exists())
-	    return;
-      if (rootFile.isDirectory()) {
-	    File[] files = rootFile.listFiles();
-	    assert files != null;
-	    for (File file : files) {
-		  if (file.isDirectory()) {
-			removeFiles(file.toPath());
-		  }
-		  Files.delete(file.toPath());
-	    }
-      }
-      Files.delete(filePath);
+      Files.walkFileTree(filePath, new PackageRemover());
 }
 
-public static void appendConfig(Path configPath, String[] aliases, Path instancePath) {
-      try {
-	    var info = new InstanceInfo(aliases, instancePath.toAbsolutePath().toString());
-	    Files.writeString(configPath, info.toString(), StandardOpenOption.APPEND);
-      } catch (IOException e) {
-	    throw new RuntimeException("Configuration file is not exists");
+private static class PackageRemover extends SimpleFileVisitor<Path> {
+      @Override
+      public FileVisitResult visitFile(Path filePath, BasicFileAttributes attrs) throws IOException {
+	    Files.delete(filePath);
+	    return FileVisitResult.CONTINUE;
       }
+      @Override
+      public FileVisitResult postVisitDirectory(Path filePath, IOException exc) throws IOException {
+	    Files.delete(filePath);
+	    return FileVisitResult.CONTINUE;
+      }
+
 }
+
 
 //OS relative functionality
-private static void linkLibraries(@NotNull Path destFile, @NotNull List<Path> dependencies) throws IOException {
-      try (BufferedWriter writer = new BufferedWriter(new FileWriter(destFile.toFile()))) {
-	    for (var path : dependencies) {
-		  writer.write(path.toAbsolutePath().toString());
+private static void linkLibraries(@NotNull Path linkPath,@NotNull File executable, @NotNull List<Path> dependencies) throws IOException {
+      if (dependencies.size() != 0){
+	    try (BufferedWriter writer = new BufferedWriter(new FileWriter(linkPath.toFile()))) {
+		  for (var path : dependencies) {
+			writer.write(path.toAbsolutePath() + "\r\n");
+		  }
 	    }
-      }
+      }//also store linkPath in executable
 }
 
 private static PayloadType convert(@NotNull String type) {
@@ -239,15 +231,15 @@ private static PayloadType convert(@NotNull String type) {
       return result;
 }
 
-private String getProgramsPath() {
-      return config.programs;
+private Path getProgramsPath() {
+      return Path.of(config.programs);
 }
 
 private String getLibrariesPath() {
       return config.libraries;
 }
 
-private String getPackagesConfig() {
+private String getPackagesConfigPath() {
       return config.infoPath;
 }
